Add "New Public Master" option to MasterChooser.
This PR adds the option of creating a New Public Master in the
MasterChooser. In order to do this, the startMaster call was moved to an
AsyncTask to prevent network activity in the UI thread.

The PR also modifies the layout of the master chooser, to accommodate
for the new option in a somewhat nicer way.