Simplify MasterChooser interface.
Now the MasterChooser UI only shows the options relevant to connecting
to a robot by default. A checkbox is added to show "advanced options"
which uncovers the new master and new private master options.

A checkbox is used instead of an on/off switch because the switch needs
a higher API level than the currently used.