List all network interfaces in master chooser
A list of detected network interfaces can be found in the advanced
options segment of the master chooser UI. Selecting an interface allows
the user to configure the ROS Hostname. If no interface is selected then
the code fall-back to the previous behaviour.

I made a simple modifications to the pubsub example to show how the
RosHostname can be used in the applications.