Add onDoubleTap to CameraControlListener
The user of a visualization view can now use the double tap gesture and
obtain the tapped position in the camera frame.

This could be useful to use the map as a way of sensing navigation
goals by tapping.

Adds the onDoubleTap method to CameraControlListener. This breaks
previous code that does not override that method. The map_viewer
tutorial is updated in this commit.